Update the dependency and the injection logic of multiple modules
