+++ adapter/src/main/java/org/mifos/chatbot/adapter/MifosChatbotAdapterService.java
@@ -2,17 +2,22 @@ package org.mifos.chatbot.adapter;
 
 import lombok.extern.slf4j.Slf4j;
 import org.mifos.chatbot.core.IntentHandler;
+import org.mifos.chatbot.core.NLPService;
 import org.mifos.chatbot.core.model.Intent;
 import org.mifos.chatbot.core.model.MifosResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Slf4j
 @Service
 public class MifosChatbotAdapterService {
+    @Autowired
+    private NLPService openNLPService;
+
     @Autowired
     private List<IntentHandler> handlers;
 
@@ -22,6 +27,15 @@ public class MifosChatbotAdapterService {
             .filter(handler -> handler.canHandle(intent))
             .map(handler -> handler.handle(intent))
             .collect(Collectors.toList());
+    }
+
+    public List<MifosResponse> handle(String input) {
+        List<MifosResponse> results = new ArrayList<>();
+        Intent[] intents = openNLPService.recognize(input);
+        for(Intent intent : intents) {
+            results.addAll(handle(intent));
+        }
+        return results;
     }
 
     /*
