+++ adapter/src/main/java/org/mifos/chatbot/adapter/handler/BaseIntentHandler.java
@@ -4,6 +4,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.mifos.chatbot.client.ApiClient;
 import org.mifos.chatbot.core.IntentHandler;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
 
 @Slf4j
 abstract class BaseIntentHandler implements IntentHandler {
