+++ client/build.gradle
@@ -8,4 +8,7 @@ dependencies {
     compile "com.google.code.gson:gson"
     compile "io.gsonfire:gson-fire"
     compile "org.threeten:threetenbp"
+    compile 'org.springframework:spring-core'
+    compile 'org.springframework:spring-context'
+    testCompile('org.springframework.boot:spring-boot-starter-test')
 }
