+++ nlp/src/test/java/org/mifos/chatbot/nlp/OpenNLPServiceTest.java
@@ -19,7 +19,8 @@ public class OpenNLPServiceTest {
     // Intent should be specific enough to point to the RESTful API of Mifos
     @Test
     public void recognizeTest() {
-        Intent[] results = nlpService.recognize("Check my next due day");
+        Intent[] results = nlpService.recognize("not at all");
+        // TODO: Update NLP model, its detection is not accurate
         for(int i = 0; i < results.length ;i++)
             System.out.println(results[i].getKeyword());
 //        Assert.assertTrue(results[0].getKeyword().equals("Interest rate"));
