+++ protocol/src/main/java/org/mifos/chatbot/protocol/SlackBot.java
@@ -1,0 +1,54 @@
+package org.mifos.chatbot.protocol;
+
+import lombok.extern.slf4j.Slf4j;
+import me.ramswaroop.jbot.core.common.Controller;
+import me.ramswaroop.jbot.core.common.EventType;
+import me.ramswaroop.jbot.core.common.JBot;
+import me.ramswaroop.jbot.core.slack.Bot;
+import me.ramswaroop.jbot.core.slack.models.Event;
+import org.mifos.chatbot.adapter.MifosChatbotAdapterService;
+import org.mifos.chatbot.core.model.MifosResponse;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Profile;
+import org.springframework.web.socket.WebSocketSession;
+
+import java.util.List;
+
+@Slf4j
+@JBot
+@Profile("slack")
+public class SlackBot extends Bot {
+    @Value("${slackBotToken}")
+    private String slackToken;
+
+    @Autowired
+    private MifosChatbotAdapterService adapter;
+
+    @Override
+    public String getSlackToken() {
+        return slackToken;
+    }
+
+    @Override
+    public Bot getSlackBot() {
+        return this;
+    }
+
+    /**
+     * Invoked when the bot receives a direct mention (@botname: message)
+     * or a direct message. NOTE: These two event types are added by jbot
+     * to make your task easier, Slack doesn't have any direct way to
+     * determine these type of events.
+     *
+     * Comments copied from JBot example
+     *
+     * @param session
+     * @param event
+     */
+    @Controller(events = {EventType.DIRECT_MENTION, EventType.DIRECT_MESSAGE})
+    public void onReceiveDM(WebSocketSession session, Event event) {
+        List<MifosResponse> responses = adapter.handle(event.getText());
+        reply(session, event, responses.toString());
+    }
+}
