--- protocol/src/main/java/org/mifos/chatbot/protocol/slack/SlackBot.java
@@ -1,47 +1,0 @@
-package org.mifos.chatbot.protocol.slack;
-
-import lombok.extern.slf4j.Slf4j;
-import me.ramswaroop.jbot.core.common.Controller;
-import me.ramswaroop.jbot.core.common.EventType;
-import me.ramswaroop.jbot.core.common.JBot;
-import me.ramswaroop.jbot.core.slack.Bot;
-import me.ramswaroop.jbot.core.slack.models.Event;
-import me.ramswaroop.jbot.core.slack.models.Message;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Profile;
-import org.springframework.web.socket.WebSocketSession;
-
-@Slf4j
-@JBot
-@Profile("slack")
-public class SlackBot extends Bot {
-    @Value("${slackBotToken}")
-    private String slackToken;
-
-    @Override
-    public String getSlackToken() {
-        return slackToken;
-    }
-
-    @Override
-    public Bot getSlackBot() {
-        return this;
-    }
-
-    /**
-     * Invoked when the bot receives a direct mention (@botname: message)
-     * or a direct message. NOTE: These two event types are added by jbot
-     * to make your task easier, Slack doesn't have any direct way to
-     * determine these type of events.
-     *
-     * Comments copied from JBot example
-     *
-     * @param session
-     * @param event
-     */
-    @Controller(events = {EventType.DIRECT_MENTION, EventType.DIRECT_MESSAGE})
-    public void onReceiveDM(WebSocketSession session, Event event) {
-//        reply(session, event, "Hi, I am " + slackService.getCurrentUser().getName());
-        reply(session, event, new Message("Hi, this is a message!"));
-    }
-}
