+++ protocol/src/main/resources/application.properties
@@ -8,6 +8,7 @@ server.port=8080
 # slack integrations
 slackApi=https://slack.com/api
 slackBotToken=xoxb-15186592304-405342609110-7PHj7JTsvcQ3wlGrGb8XLMjW
+#slackBotToken=xoxb-15186592304-404656413348-n10AfHjSQ4NFgLzFHycW3rhX
 #slashCommandToken=<paste your slash command token>
 # replace the below url with your slack webhook url
 slackIncomingWebhookUrl=https://hooks.slack.com/services/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
