+++ protocol/src/test/java/org/mifos/chatbot/protocol/SlackIntegrationTest.java
@@ -1,7 +1,6 @@
 package org.mifos.chatbot.protocol;
 
 import org.junit.Test;
-import org.mifos.chatbot.protocol.slack.SlackIntegration;
 
 public class SlackIntegrationTest {
     SlackIntegration slackIntegrator;
